auth/requirements.txt
fastapi~=0.115.12
uvicorn~=0.34.2
sqlalchemy~=2.0.40
psycopg2-binary
python-jose~=3.4.0
python-multipart
passlib[bcrypt]~=1.7.4
pydantic[email]~=2.11.3
pydantic[email]
bcrypt==4.0.1
python-dotenv

starlette~=0.46.2
alembic~=1.15.2